+++ src/main/java/io/github/devatherock/artifactory/service/DockerBadgeService.java
@@ -18,6 +18,8 @@ import lombok.extern.slf4j.Slf4j;
 
 import javax.inject.Singleton;
 import java.text.DecimalFormat;
+import java.util.HashMap;
+import java.util.Map;
 
 @Slf4j
 @Blocking
@@ -27,6 +29,16 @@ public class DockerBadgeService {
     private static final double BYTES_IN_MB = 1024 * 1024;
     private static final String FILE_NAME_MANIFEST = "/manifest.json";
     private static final String HDR_API_KEY = "X-JFrog-Art-Api";
+    private static final double PULLS_REDUCER = 1000;
+    private static final int MAX_REDUCTIONS = 3;
+    private static final Map<Integer, String> PULLS_SUFFIX = new HashMap<>();
+
+    static {
+        PULLS_SUFFIX.put(0, "");
+        PULLS_SUFFIX.put(1, "k");
+        PULLS_SUFFIX.put(2, "M");
+        PULLS_SUFFIX.put(3, "G");
+    }
 
     private final BlockingHttpClient artifactoryClient;
     private final BadgeGenerator badgeGenerator;
@@ -46,7 +58,8 @@ public class DockerBadgeService {
                     .get() / BYTES_IN_MB;
 
             LOGGER.info("Size of {}/{}: {} MB", packageName, tag, imageSize);
-            return badgeGenerator.generateBadge(badgeLabel, String.format("%s MB", formatDecimal(imageSize)));
+            return badgeGenerator.generateBadge(badgeLabel, String.format("%s MB",
+                    formatDecimal(imageSize, "0.##")));
         } else {
             return generateNotFoundBadge(badgeLabel);
         }
@@ -89,16 +102,29 @@ public class DockerBadgeService {
                 }
             }
             LOGGER.info("Download count of {}: {}", packageName, downloadCount);
-            return badgeGenerator.generateBadge(badgeLabel, String.valueOf(downloadCount));
+            return badgeGenerator.generateBadge(badgeLabel, formatDownloadCount(downloadCount));
         } else {
             return generateNotFoundBadge(badgeLabel);
         }
     }
 
+    /**
+     * Generates a not found badge
+     *
+     * @param badgeLabel the badge label
+     * @return a not found badge
+     */
     private String generateNotFoundBadge(String badgeLabel) {
         return badgeGenerator.generateBadge(badgeLabel, "Not Found");
     }
 
+    /**
+     * Reads a docker {@code manifest.json} file
+     *
+     * @param packageName the docker image name
+     * @param tag         the docker image tag
+     * @return a {@link DockerManifest}
+     */
     private DockerManifest readManifest(String packageName, String tag) {
         String fullPackageName = packageName + "/" + tag;
         HttpRequest<Object> manifestRequest = HttpRequest.create(HttpMethod.GET, artifactoryConfig.getUrlPrefix()
@@ -109,12 +135,31 @@ public class DockerBadgeService {
     /**
      * Formats a decimal number into a string
      *
-     * @param inputDecimal
+     * @param inputDecimal the decimal number to format
+     * @param format       the pattern to which to format to
      * @return a formatted string
      */
-    private String formatDecimal(double inputDecimal) {
-        DecimalFormat decimalFormat = new DecimalFormat("0.##");
+    private String formatDecimal(double inputDecimal, String format) {
+        DecimalFormat decimalFormat = new DecimalFormat(format);
         decimalFormat.setDecimalSeparatorAlwaysShown(false);
         return decimalFormat.format(inputDecimal);
+    }
+
+    /**
+     * Formats the download count with a suffix
+     *
+     * @param downloadCount number of downloads
+     * @return a formatted string
+     */
+    private String formatDownloadCount(long downloadCount) {
+        double reducedCount = downloadCount;
+        int reductions = 0;
+
+        while (reducedCount > PULLS_REDUCER && reductions < MAX_REDUCTIONS) {
+            reductions++;
+            reducedCount = reducedCount / PULLS_REDUCER;
+        }
+
+        return formatDecimal(reducedCount, "0.#") + PULLS_SUFFIX.get(reductions);
     }
 }