+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Set the number of threads on the IO pool to `200`, to match spring boot's tomcat
 - Upgraded JRE base image to `11.0.8` from `11.0.1` to fix SSL errors with TLS v1.3
 - [#10](https://github.com/devatherock/artifactory-badge/issues/10): Hid decimal point when image size is a whole number
+- [#11](https://github.com/devatherock/artifactory-badge/issues/11): Shortened pulls count
 
 ### Removed
 - [#13](https://github.com/devatherock/artifactory-badge/issues/13): Apache HTTP Client